[src] widget/flexbox: create stub
This trait approach doesn't work either.

`Widget` (i.e. `Responsive`) is trivial and object safe.  The problem
comes from `Reactive` and `Command`.  The latter isn't object safe
because of a generic writer.  This can be bypassed by bringing back the
`View` trait and have it be a copy of `Command`, but with a defined
writer type.  As long as it implements `fmt::Write`, crossterm's
commands can be used on it.  It's hacky, but it could work.

`Reactive` is a bigger problem.  It requires each type to have their own
`CustomMsg`.  And you can't roll up dynamic `Widget` objects, each with
their own message type.

In general, Rust isn't conducive to this web-inspired design.  `Flexbox`
needs to propagate messages.  But it's impossible to define an enum with
all of the child messages, since `Flexbox::from` is a runtime function.

In summary, `Responsive` is fine, `View` is tricky, `Reactive` is
impossible in its current form.