[src] widget: add an `Area` argument to `on_click`
This way widgets can determine if the event is relevant or not.  In
theory, it only needs to be done for the top element, which can be
handled by the event loop.  But then each container widget would have to
tweak the coordinates inside `MouseEvent`, which is unergonomic.  I can
simply make a method on `Area` instead.